Induction machine

Induction machine model for detailed representation of electrical loads.

Disclaimer: This model has only been tested for the power flow problem with ACP formulation.

Parameters

Set of parameters used to model the induction machine as defined in the input data

name	symb.	unit	type	default	definition

index	$im$	-	Int	-	unique index of the induction machine

im_bus	$i$	-	Int	-	unique index of the bus to which the induction machine is connected to

P_ag	$P_{ag}$	p.u.	Real	-	Starting value for active power of induction machine - positive for consumption

Q_ag	$Q_{ag}$	p.u.	Real	-	Starting value for active power of induction machine - positive for consumption

Pacmin	$\underline{P_{im}}$	p.u.	Real	-	minimum stable operating power of the induction machine

Pacmax	$\overline{P_{im}}$	p.u.	Real	-	maximum power rating of the induction machine

Pacrated	$P_{im}$	p.u.	Real	-	minimum reactive power of the induction machine

status	$\delta_{im}$	-	Int	-	Status indicator of the induction machine

x_m	$x_{m}$	p.u.	Real	-	Magnetizing inductance of induction machine

x_sl	$x_{sl}$	p.u.	Real	-	Stator leakage inductance of induction machine

x_rl	$x_{rl}$	p.u.	Real	-	Rotor leakage inductance of induction machine

r_s	$r_{s}$	p.u.	Real	-	Stator resistance of induction machine

r_r	$r_{r}$	p.u.	Real	-	Rotor resistance of induction machine

torque	-	-	-	-	Torque model parameters of induction machine

Torque parameter

The torque model is implemented as $T(\omega) = T_0 * (A*\omega^m+B \omega + C)$. This representation includes both quadratic models (m=2) or power functions (B=C=0) (see Kundur).

Name	Symbol	Unit	Type	Default	Description

T_0	$T_0$	p.u.	Real	-	Per-unit torque scaling factor (base torque is approximately equal to the system base power)

A	$A$	$\mathrm{s}^2/\mathrm{rad}^2$	Real	-	Quadratic coefficient of the mechanical torque characteristic (load component typical for centrifugal devices or aerodynamic drag)

B	$B$	$\mathrm{s}/\mathrm{rad}$	Real	-	Linear coefficient of the mechanical torque characteristic (friction load component)

C	$C$	-	Real	-	Constant coefficient of the mechanical torque characteristic (static load component)

m	$m$	-	Real	-	Mechanical torque exponent describing the load type (see Kundur)

Variables

The main optimisation variables of interest are:

name	symb.	unit	formulation	definition

pg	$P_{g}$	p.u.	ACP	Active power of induction machine g

qg	$Q_{g}$	p.u.	ACP	Reactive power of induction machine g

Constraints

See Kundur or Van Cutsem for detailed derivation of equations.

Stator constrains

PowerModelsACDC.constraint_im_stator — Function

Im stator constraints (Based on transformer formulation)

\[p_{im,s,fr} = g v_{m,fr}^{2} - g v_{m,fr} v_{m,to}\cos\left(v_{a,fr}-v_{a,to}\right) - b v_{m,fr} v_{m,to}\sin\left(v_{a,fr}-v_{a,to}\right)\]

\[q_{im,s,fr} = - b v_{m,fr}^{2} + b v_{m,fr} v_{m,to}\cos\left(v_{a,fr}-v_{a,to}\right) - g v_{m,fr} v_{m,to}\sin\left(v_{a,fr}-v_{a,to}\right)\]

\[p_{im,s,to} = g v_{m,to}^{2} - g v_{m,to} v_{m,fr}\cos\left(v_{a,to}-v_{a,fr}\right) - b v_{m,to} v_{m,fr}\sin\left(v_{a,to}-v_{a,fr}\right)\]

\[q_{im,s,to} = - b v_{m,to}^{2} + b v_{m,to} v_{m,fr}\cos\left(v_{a,to}-v_{a,fr}\right) - g v_{m,to} v_{m,fr}\sin\left(v_{a,to}-v_{a,fr}\right)\]

Rotor inductance constraints

PowerModelsACDC.constraint_im_rotor_inductance — Function

IM rotor inductance constraints

\[-p_{im,ag} = g_c v_{m,ag}^{2} - g_c v_{m,ag} v_{m,m}\cos\left(v_{a,ag}-v_{a,m}\right) - b_c v_{m,ag} v_{m,m}\sin\left(v_{a,ag}-v_{a,m}\right)\]

\[-q_{im,ag} = - b_c v_{m,ag}^{2} + b_c v_{m,ag} v_{m,m}\cos\left(v_{a,ag}-v_{a,m}\right) - g_c v_{m,ag} v_{m,m}\sin\left(v_{a,ag}-v_{a,m}\right)\]

\[p_{im,ri,f} = g_c v_{m,m}^{2} - g_c v_{m,m} v_{m,ag}\cos\left(v_{a,m}-v_{a,ag}\right) - b_c v_{m,m} v_{m,ag}\sin\left(v_{a,m}-v_{a,ag}\right)\]

\[q_{im,ri,f} = - b_c v_{m,m}^{2} + b_c v_{m,m} v_{m,ag}\cos\left(v_{a,m}-v_{a,ag}\right) - g_c v_{m,m} v_{m,ag}\sin\left(v_{a,m}-v_{a,ag}\right)\]

Magnetistation branch constraints

PowerModelsACDC.constraint_im_magnetisation — Function

IM magnetisation constraint

\[p_{im,ri,f}+p_{im,s,to}=0\]

\[q_{im,ri,f}+q_{im,s,to}+b_v v_{m,m}^{2}=0\]

Slip constraints

PowerModelsACDC.constraint_im_slip — Function

IM slip constraints (slip from balance air-gap power with mechanical torque, see Van Cutsem Voltage stability)

\[T_0\left(A(1-s)^m+B(1-s)+C\right) = \frac{v_{m,ag}^{2}s}{r_r}\]

\[p_{im,ag} = \frac{v_{m,ag}^{2}s}{r_r}\]

\[q_{im,ag}=0\]