Variables

Functions

We provide the following methods to provide a compositional approach for defining common variables used in power flow models. These methods should always be defined over "GenericPowerModel", from the base PowerModels.jl.

PowerModelsACDC.variable_acside_current — Method

variable: iconv_ac[j] and iconv_ac_sq[j] for j in convdc for ACR formulation

PowerModelsACDC.variable_acside_current — Method

variable: iconv_ac[j] for j in convdc

PowerModelsACDC.variable_acside_current — Method

variable: iconv_ac[j] and iconv_ac_sq[j] for j in convdc

PowerModelsACDC.variable_acside_current_ne — Method

variable: iconv_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_acside_current_ne — Method

variable: iconv_ac_ne[j] and iconv_ac_sq_ne[j] for j in candidate convdc

PowerModelsACDC.variable_conv_reactor_active_power_from — Method

variable: pconv_pr_from[j] for j in convdc

PowerModelsACDC.variable_conv_reactor_active_power_from_ne — Method

variable: pconv_pr_from_ne[j] for j in candidate convdc

PowerModelsACDC.variable_conv_reactor_current_sqr — Method

variable: irc_sq[j] for j in convdc

PowerModelsACDC.variable_conv_reactor_current_sqr_ne — Method

variable: irc_sq_ne[j] for j in candidate convdc

PowerModelsACDC.variable_conv_reactor_reactive_power_from — Method

variable: qconv_pr_from[j] for j in convdc

PowerModelsACDC.variable_conv_reactor_reactive_power_from_ne — Method

variable: qconv_pr_from_ne[j] for j in candidate convdc

PowerModelsACDC.variable_conv_transformer_active_power_to — Method

variable: pconv_grid_ac_to[j] for j in convdc

PowerModelsACDC.variable_conv_transformer_active_power_to_ne — Method

variable: pconv_grid_ac_to_ne[j] for j in candidate convdc

PowerModelsACDC.variable_conv_transformer_current_sqr — Method

variable: itf_sq[j] for j in convdc

PowerModelsACDC.variable_conv_transformer_current_sqr_ne — Method

variable: itf_sq_ne[j] for j in candidate convdc

PowerModelsACDC.variable_conv_transformer_reactive_power_to — Method

variable: qconv_grid_ac_to[j] for j in convdc

PowerModelsACDC.variable_conv_transformer_reactive_power_to_ne — Method

variable: qconv_grid_ac_to_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_active_power — Method

variable: pconv_ac[j] for j in convdc

PowerModelsACDC.variable_converter_active_power_ne — Method

variable: pconv_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_filter_voltage_angle — Method

variable: vaf[j] for j in convdc

PowerModelsACDC.variable_converter_filter_voltage_angle_ne — Method

variable: vaf_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_filter_voltage_cross_products — Method

variable: wrf_ac[j] and wif_ac for j in convdc

PowerModelsACDC.variable_converter_filter_voltage_cross_products_ne — Method

variable: wrf_ac_ne[j] and wif_ac for j in candidate convdc

PowerModelsACDC.variable_converter_filter_voltage_imaginary — Method

imaginary part of the voltage variable vif[j] for j in convdc

PowerModelsACDC.variable_converter_filter_voltage_magnitude — Method

variable: vmf[j] for j in convdc

PowerModelsACDC.variable_converter_filter_voltage_magnitude_ne — Method

variable: vmf_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_filter_voltage_magnitude_sqr — Method

variable: wf_ac for j in convdc

PowerModelsACDC.variable_converter_filter_voltage_magnitude_sqr_ne — Method

variable: wf_ac_ne for j in candidate convdc

PowerModelsACDC.variable_converter_filter_voltage_real — Method

real part of the voltage variable vrf[j] for j in convdc

PowerModelsACDC.variable_converter_firing_angle — Method

variable: pconv_dc[j] for j in convdc

PowerModelsACDC.variable_converter_firing_angle_ne — Method

variable: pconv_dc_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_internal_voltage_angle — Method

variable: vac[j] for j in convdc

PowerModelsACDC.variable_converter_internal_voltage_angle_ne — Method

variable: vac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_internal_voltage_cross_products — Method

variable: wrc_ac[j] and wic_ac[j] for j in convdc

PowerModelsACDC.variable_converter_internal_voltage_cross_products_ne — Method

variable: wrc_ac_ne[j] and wic_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_internal_voltage_imaginary — Method

imaginary part of the voltage variable vic[j] for j in convdc

PowerModelsACDC.variable_converter_internal_voltage_magnitude — Method

variable: vmc[j] for j in convdc

PowerModelsACDC.variable_converter_internal_voltage_magnitude_ne — Method

variable: vmc_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_internal_voltage_magnitude_sqr — Method

variable: wc_ac[j] for j in convdc

PowerModelsACDC.variable_converter_internal_voltage_magnitude_sqr_ne — Method

variable: wc_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_internal_voltage_real — Method

real part of the voltage variable vrc[j] for j in convdc

PowerModelsACDC.variable_converter_ne — Method

variable: 0 <= convdc_ne[c] <= 1 for c in `candidate converters

PowerModelsACDC.variable_converter_reactive_power — Method

variable: qconv_ac[j] for j in convdc

PowerModelsACDC.variable_converter_reactive_power_ne — Method

variable: qconv_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_to_grid_active_power — Method

variable: pconv_grid_ac[j] for j in convdc

PowerModelsACDC.variable_converter_to_grid_active_power_ne — Method

variable: pconv_grid_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_converter_to_grid_reactive_power — Method

variable: qconv_grid_ac[j] for j in convdc

PowerModelsACDC.variable_converter_to_grid_reactive_power_ne — Method

variable: qconv_grid_ac_ne[j] for j in candidate convdc

PowerModelsACDC.variable_dc_converter — Method

All converter variables

PowerModelsACDC.variable_dc_converter_ne — Method

All converter variables

PowerModelsACDC.variable_dcside_power — Method

variable: pconv_dc[j] for j in convdc

PowerModelsACDC.variable_dcside_power_ne — Method

variable: pconv_dc_ne[j] for j in candidate convdc

PowerModelsACDC.variable_active_dcbranch_flow — Method

variable: p_dcgrid[l,i,j] for (l,i,j) in arcs_dcgrid

PowerModelsACDC.variable_active_dcbranch_flow_ne — Method

variable: p_dcgrid[l,i,j] for (l,i,j) in arcs_dcgrid

PowerModelsACDC.variable_branch_ne — Method

variable: 0 <= convdc_ne[c] <= 1 for c in `candidate converters

PowerModelsACDC.variable_dcbranch_current_sqr — Method

variable: ccm_dcgrid[l] for (l) in branchdc

PowerModelsACDC.variable_dcbranch_current_sqr_ne — Method

variable: ccm_dcgrid[l] for (l) in branchdc

PowerModelsACDC.variable_dcgrid_voltage_magnitude — Method

variable: vdcm[i] for i in dcbuses

PowerModelsACDC.variable_dcgrid_voltage_magnitude_ne — Method

variable: vdcm[i] for i in dcbuses

PowerModelsACDC.variable_dcgrid_voltage_magnitude_sqr — Method

variable: vdcm[i] for i in dcbuses

PowerModelsACDC.variable_dcgrid_voltage_magnitude_sqr_ne — Method

variable: vdcm[i] for i in dcbuses