Objective

OPF Objective

Minimisation of active power geneation costs

TNEP Objectives

Minimisation of consturction cost for HVDC branches and converters

Minimisation of consturction cost for HVDC branches and converters and AC branches (acdc_tnep)

Minimisation of consturction cost for HVDC branches and converters and generation costs

Minimisation of consturction cost for HVDC branches and converters and generation costs and AC branches(acdc_tnep)

PowerModelsACDC.objective_min_fuel_cost — Function

Minimum fuel cost objective using lienar and quadratic terms or piecewise linear functions, as defined in Matpower

PowerModelsACDC.objective_min_cost — Function

Objective consisting of generation cost + investment costs of HVDC lines and converters

PowerModelsACDC.objective_min_cost_acdc — Function

Objective consisting of generation cost + investment costs of HVDC lines and converters + investment costs of AC lines